requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
Flask-Bcrypt
email_validator
Flask-SocketIO
eventlet
Pillow